refactor(car): car 패키지를 domain 패키지로 이동
