+++ src/main/java/racinggame/domain/car/RacingCar.java
@@ -1,4 +1,4 @@
-package racinggame.car;
+package racinggame.domain.car;
 
 public class RacingCar {
