+++ src/main/java/racinggame/domain/car/RandomMovableStrategy.java
@@ -1,4 +1,4 @@
-package racinggame.car;
+package racinggame.domain.car;
 
 import java.util.Random;
