+++ src/test/java/racinggame/domain/car/RacingCarTest.java
@@ -1,4 +1,4 @@
-package racinggame.car;
+package racinggame.domain.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
